+++ packages/cli/src/capture/index.ts
@@ -22,6 +22,19 @@ import {
   startCdpAnimationCapture,
   collectAnimationCatalog,
 } from "./animationCataloger.js";
+import {
+  saveLottieAnimations,
+  renderLottiePreviews,
+  captureVideoManifest,
+} from "./mediaCapture.js";
+import type { DiscoveredLottie } from "./mediaCapture.js";
+import {
+  detectLibraries,
+  extractVisibleText,
+  captionImagesWithGemini,
+  generateAssetDescriptions,
+} from "./contentExtractor.js";
+import { loadEnvFile, generateProjectScaffold } from "./scaffolding.js";
 import type { CaptureOptions, CaptureResult } from "./types.js";
 
 export type { CaptureOptions, CaptureResult } from "./types.js";
@@ -47,35 +60,7 @@ export async function captureWebsite(
   };
 
   // Load .env file from repo root if it exists (for GEMINI_API_KEY, etc.)
-  try {
-    const { readFileSync: readEnv } = await import("node:fs");
-    const { resolve: resolvePath } = await import("node:path");
-    // Walk up from outputDir to find .env in repo root
-    let dir = resolvePath(outputDir);
-    for (let i = 0; i < 5; i++) {
-      const envPath = resolvePath(dir, ".env");
-      try {
-        const envContent = readEnv(envPath, "utf-8");
-        for (const line of envContent.split("\n")) {
-          const trimmed = line.trim();
-          if (!trimmed || trimmed.startsWith("#")) continue;
-          const eq = trimmed.indexOf("=");
-          if (eq === -1) continue;
-          const key = trimmed.slice(0, eq).trim();
-          const val = trimmed
-            .slice(eq + 1)
-            .trim()
-            .replace(/^["']|["']$/g, "");
-          if (!process.env[key]) process.env[key] = val;
-        }
-        break;
-      } catch {
-        dir = resolvePath(dir, "..");
-      }
-    }
-  } catch {
-    /* .env loading is best-effort */
-  }
+  loadEnvFile(outputDir);
 
   // Create output directories
   mkdirSync(join(outputDir, "extracted"), { recursive: true });
@@ -156,12 +141,7 @@ export async function captureWebsite(
     `);
 
     // Intercept network responses to detect Lottie JSON files
-    const discoveredLotties: Array<{
-      url: string;
-      data?: unknown;
-      dimensions?: { w: number; h: number };
-      frameRate?: number;
-    }> = [];
+    const discoveredLotties: DiscoveredLottie[] = [];
     page1.on("response", async (response) => {
       try {
         const responseUrl = response.url();
@@ -204,14 +184,23 @@ export async function captureWebsite(
     await new Promise((r) => setTimeout(r, settleTime));
 
     // Check if the page loaded real content or an anti-bot challenge
+    // Use structural detection (DOM elements + cookies), not text regex matching —
+    // text matching causes false positives on sites that mention "blocked" or "verify" in copy
     const pageContentCheck = (await page1.evaluate(`(() => {
       var text = (document.body.innerText || "").trim();
       var title = document.title || "";
-      var isChallenged = /cloudflare|just a moment|checking your browser|access denied|blocked|captcha|verify you are human/i.test(text + " " + title);
-      return { textLength: text.length, title: title, isChallenged: isChallenged };
-    })()`)) as { textLength: number; title: string; isChallenged: boolean };
+      // Structural: Cloudflare Turnstile widget or challenge iframe
+      var hasCfTurnstile = !!document.querySelector('.cf-turnstile, [data-sitekey], iframe[src*="challenges.cloudflare.com"], #challenge-running, #challenge-form');
+      // Structural: page is almost empty (challenge pages have minimal DOM)
+      var bodyChildCount = document.body.children.length;
+      var isMinimalDom = bodyChildCount <= 5 && text.length < 500;
+      // Title-based: only check title on near-empty pages
+      var hasChallengeTitle = isMinimalDom && /just a moment|attention required|access denied/i.test(title);
+      var isChallenged = hasCfTurnstile || hasChallengeTitle;
+      return { textLength: text.length, title: title, isChallenged: isChallenged, bodyChildCount: bodyChildCount };
+    })()`)) as { textLength: number; title: string; isChallenged: boolean; bodyChildCount: number };
 
-    if (pageContentCheck.isChallenged || pageContentCheck.textLength < 200) {
+    if (pageContentCheck.isChallenged || pageContentCheck.textLength < 100) {
       const reason = pageContentCheck.isChallenged
         ? "Anti-bot protection detected (Cloudflare challenge or similar)"
         : "Page has very little text content (" +
@@ -219,12 +208,6 @@ export async function captureWebsite(
           " chars) — may be blocked or a client-rendered SPA that needs more time";
       warnings.push(reason);
       progress("warn", reason);
-      // Write a BLOCKED.md file so the user/agent knows what happened
-      writeFileSync(
-        join(outputDir, "BLOCKED.md"),
-        `# Capture Warning\n\n${reason}\n\nThe page at ${url} may not have been captured correctly.\n\n## What to try\n\n- Re-run with a longer timeout: \`--timeout 60000\`\n- The site may require authentication or block headless browsers\n- Try capturing from a different URL on the same domain\n`,
-        "utf-8",
-      );
     }
 
     // Scroll through page to trigger lazy-loaded images and Lottie animations
@@ -275,157 +258,16 @@ export async function captureWebsite(
     if (discoveredLotties.length > 0) {
       const lottieDir = join(outputDir, "assets", "lottie");
       mkdirSync(lottieDir, { recursive: true });
-      let savedCount = 0;
-      const savedHashes = new Set<string>(); // Deduplicate by content
-
-      for (let li = 0; li < discoveredLotties.length && li < 10; li++) {
-        const lottieItem = discoveredLotties[li]!;
-        try {
-          let jsonData: string | undefined;
-
-          if (lottieItem.data) {
-            // Already have the JSON data from network interception
-            jsonData = JSON.stringify(lottieItem.data);
-          } else if (lottieItem.url) {
-            // Download the file
-            const res = await fetch(lottieItem.url, {
-              signal: AbortSignal.timeout(10000),
-              headers: { "User-Agent": "HyperFrames/1.0" },
-            });
-            if (!res.ok) continue;
-            const buf = Buffer.from(await res.arrayBuffer());
-
-            if (lottieItem.url.endsWith(".lottie")) {
-              // dotLottie is a ZIP — extract the animation JSON
-              try {
-                const AdmZip = (await import("adm-zip")).default;
-                const zip = new AdmZip(buf);
-                const entries = zip.getEntries();
-                // Look for animation JSON in both v1 (animations/) and v2 (a/) paths
-                const animEntry = entries.find(
-                  (e) =>
-                    (e.entryName.startsWith("a/") || e.entryName.startsWith("animations/")) &&
-                    e.entryName.endsWith(".json"),
-                );
-                if (animEntry) {
-                  jsonData = animEntry.getData().toString("utf-8");
-                }
-              } catch {
-                // adm-zip not available or extraction failed — save raw .lottie
-                const hash = buf.toString("base64").slice(0, 100);
-                if (savedHashes.has(hash)) continue;
-                savedHashes.add(hash);
-                writeFileSync(join(lottieDir, `animation-${savedCount}.lottie`), buf);
-                savedCount++;
-                continue;
-              }
-            } else {
-              // Plain JSON file
-              jsonData = buf.toString("utf-8");
-            }
-          }
-
-          if (jsonData) {
-            // Deduplicate by content hash (first 100 chars of stringified JSON)
-            const hash = jsonData.slice(0, 200);
-            if (savedHashes.has(hash)) continue;
-            savedHashes.add(hash);
-
-            // Validate it's actually Lottie
-            try {
-              const parsed = JSON.parse(jsonData);
-              if (!parsed.layers || !parsed.w) continue;
-            } catch {
-              continue;
-            }
-
-            writeFileSync(join(lottieDir, `animation-${savedCount}.json`), jsonData, "utf-8");
-            savedCount++;
-          }
-        } catch {
-          /* skip */
-        }
-      }
+      const savedCount = await saveLottieAnimations(discoveredLotties, lottieDir);
       // Generate manifest + preview thumbnails so the agent can SEE what each animation is
       if (savedCount > 0) {
-        const manifest: Array<{
-          file: string;
-          preview: string;
-          name: string;
-          width: number;
-          height: number;
-          duration: number;
-          frameRate: number;
-          layers: number;
-        }> = [];
-        const { readdirSync, readFileSync: readFs } = await import("node:fs");
-        const previewDir = join(lottieDir, "previews");
-        mkdirSync(previewDir, { recursive: true });
-
-        for (const file of readdirSync(lottieDir)) {
-          if (!file.endsWith(".json")) continue;
-          try {
-            const raw = JSON.parse(readFs(join(lottieDir, file), "utf-8"));
-            const fr = raw.fr || 30;
-            const dur = ((raw.op || 0) - (raw.ip || 0)) / fr;
-            const previewName = file.replace(".json", "-preview.png");
-
-            // Render a mid-frame thumbnail using Puppeteer + lottie-web
-            try {
-              const previewPage = await chromeBrowser.newPage();
-              await previewPage.setViewport({ width: 400, height: 400 });
-              const animJson = readFs(join(lottieDir, file), "utf-8");
-              const midFrame = Math.floor(((raw.op || 0) - (raw.ip || 0)) * 0.3);
-              await previewPage.setContent(
-                `<!DOCTYPE html>
-<html><head>
-<script src="https://cdnjs.cloudflare.com/ajax/libs/lottie-web/5.12.2/lottie.min.js"></script>
-<style>*{margin:0;padding:0;background:transparent}#c{width:400px;height:400px}</style>
-</head><body><div id="c"></div><script>
-var a=lottie.loadAnimation({container:document.getElementById('c'),renderer:'svg',loop:false,autoplay:false,animationData:${animJson.replace(/</g, "\\u003c")}});
-a.addEventListener('DOMLoaded',function(){a.goToAndStop(${midFrame},true);window.__READY=true});
-</script></body></html>`,
-                { waitUntil: "networkidle0", timeout: 10000 },
-              );
-              await previewPage
-                .waitForFunction(() => (window as any).__READY === true, { timeout: 5000 })
-                .catch(() => {});
-              await previewPage.screenshot({
-                path: join(previewDir, previewName),
-                type: "png",
-                omitBackground: true,
-              });
-              await previewPage.close();
-            } catch {
-              /* preview rendering failed — non-critical */
-            }
-
-            manifest.push({
-              file: `assets/lottie/${file}`,
-              preview: `assets/lottie/previews/${previewName}`,
-              name: raw.nm || file,
-              width: raw.w || 0,
-              height: raw.h || 0,
-              duration: Math.round(dur * 10) / 10,
-              frameRate: fr,
-              layers: (raw.layers || []).length,
-            });
-          } catch {
-            /* skip */
-          }
-        }
-        if (manifest.length > 0) {
-          writeFileSync(
-            join(outputDir, "extracted", "lottie-manifest.json"),
-            JSON.stringify(manifest, null, 2),
-            "utf-8",
-          );
-        }
+        await renderLottiePreviews(chromeBrowser, lottieDir, outputDir);
         progress("lottie", `${savedCount} Lottie animation(s) saved`);
       }
     }
 
     // Save captured WebGL shaders (useful context for shader transitions + library detection)
+    let capturedShaders: Array<{ type: string; source: string }> | undefined;
     try {
       const shaders = await page1.evaluate(`window.__capturedShaders || []`);
       if (Array.isArray(shaders) && shaders.length > 0) {
@@ -435,6 +277,7 @@ a.addEventListener('DOMLoaded',function(){a.goToAndStop(${midFrame},true);window
           seen.add(s.source);
           return true;
         });
+        capturedShaders = unique;
         writeFileSync(
           join(outputDir, "extracted", "shaders.json"),
           JSON.stringify(unique, null, 2),
@@ -526,268 +369,16 @@ a.addEventListener('DOMLoaded',function(){a.goToAndStop(${midFrame},true);window
     // Generate video manifest — screenshot each <video> element + extract surrounding context
     // so Claude Code can SEE what each video shows and WHERE it was used on the page.
     try {
-      const videoElements = (await page1.evaluate(`(() => {
-        var videos = Array.from(document.querySelectorAll('video'));
-        return videos.map(function(v) {
-          var src = v.src || v.currentSrc || (v.querySelector('source') ? v.querySelector('source').src : '');
-          if (!src || !src.startsWith('http')) return null;
-
-          // Get bounding box for screenshot
-          var rect = v.getBoundingClientRect();
-          if (rect.width < 10 || rect.height < 10) return null;
-
-          // Nearest heading above the video
-          var heading = '';
-          var el = v;
-          for (var i = 0; i < 8; i++) {
-            el = el.parentElement;
-            if (!el) break;
-            var h = el.querySelector('h1,h2,h3,h4');
-            if (h) { heading = h.textContent.trim().slice(0, 100); break; }
-          }
-
-          // Nearest paragraph/caption text
-          var caption = '';
-          el = v;
-          for (var j = 0; j < 5; j++) {
-            el = el.parentElement;
-            if (!el) break;
-            var p = el.querySelector('p,figcaption,[class*="caption"],[class*="desc"]');
-            if (p) { caption = p.textContent.trim().slice(0, 200); break; }
-          }
-
-          // aria-label on video or wrapper
-          var ariaLabel = v.getAttribute('aria-label') || v.getAttribute('title') || '';
-          var wrapper = v.parentElement;
-          if (!ariaLabel && wrapper) ariaLabel = wrapper.getAttribute('aria-label') || '';
-
-          return {
-            src: src,
-            width: Math.round(rect.width),
-            height: Math.round(rect.height),
-            top: Math.round(rect.top),
-            left: Math.round(rect.left),
-            heading: heading,
-            caption: caption,
-            ariaLabel: ariaLabel,
-            filename: src.split('/').pop().split('?')[0],
-          };
-        }).filter(Boolean);
-      })`)) as Array<{
-        src: string;
-        width: number;
-        height: number;
-        top: number;
-        left: number;
-        heading: string;
-        caption: string;
-        ariaLabel: string;
-        filename: string;
-      }>;
-
-      // Deduplicate by src
-      const seenSrcs = new Set<string>();
-      const uniqueVideos = videoElements.filter((v) => {
-        if (seenSrcs.has(v.src)) return false;
-        seenSrcs.add(v.src);
-        return true;
-      });
-
-      if (uniqueVideos.length > 0) {
-        const videoManifestDir = join(outputDir, "assets", "videos");
-        mkdirSync(videoManifestDir, { recursive: true });
-        const previewDir = join(videoManifestDir, "previews");
-        mkdirSync(previewDir, { recursive: true });
-
-        const videoManifest: Array<{
-          index: number;
-          url: string;
-          filename: string;
-          width: number;
-          height: number;
-          heading: string;
-          caption: string;
-          ariaLabel: string;
-          preview: string;
-        }> = [];
-
-        for (let vi = 0; vi < uniqueVideos.length && vi < 20; vi++) {
-          const v = uniqueVideos[vi]!;
-          const previewName = `video-${vi}-preview.png`;
-          const previewPath = join(previewDir, previewName);
-
-          // Screenshot the video element to get a visible frame
-          try {
-            // Scroll to the video element so it's in the viewport
-            await page1.evaluate(`window.scrollTo(0, ${Math.max(0, v.top - 100)})`);
-            await new Promise((r) => setTimeout(r, 300));
-            // Re-measure position after scroll (layout may have shifted)
-            const rect = (await page1.evaluate((fn) => {
-              const vid = [...document.querySelectorAll("video")].find((x) =>
-                (x.src || x.currentSrc || "").includes(fn),
-              );
-              if (!vid) return null;
-              // Seek to 0.1s and wait for a frame to decode
-              vid.currentTime = 0.1;
-              return vid.getBoundingClientRect().toJSON();
-            }, v.filename)) as { x: number; y: number; width: number; height: number } | null;
-            if (!rect || rect.width < 10) continue;
-            await new Promise((r) => setTimeout(r, 200)); // let decoder settle
-            await page1.screenshot({
-              path: previewPath,
-              clip: {
-                x: Math.max(0, rect.x),
-                y: Math.max(0, rect.y),
-                width: Math.min(rect.width, 1920),
-                height: Math.min(rect.height, 1080),
-              },
-            });
-          } catch {
-            /* preview failed — non-critical */
-          }
-
-          videoManifest.push({
-            index: vi,
-            url: v.src,
-            filename: v.filename,
-            width: v.width,
-            height: v.height,
-            heading: v.heading,
-            caption: v.caption,
-            ariaLabel: v.ariaLabel,
-            preview: `assets/videos/previews/${previewName}`,
-          });
-        }
-
-        if (videoManifest.length > 0) {
-          writeFileSync(
-            join(outputDir, "extracted", "video-manifest.json"),
-            JSON.stringify(videoManifest, null, 2),
-            "utf-8",
-          );
-          progress("design", `${videoManifest.length} video previews captured`);
-        }
-      }
+      await captureVideoManifest(page1, outputDir, progress);
     } catch {
       /* non-blocking — video manifest is best-effort */
     }
 
-    // Detect JS libraries via globals, DOM fingerprints, and script URLs
-    let detectedLibraries: string[] = [];
-    try {
-      detectedLibraries = (await page1.evaluate(`(() => {
-        var libs = [];
-        function add(name) { if (libs.indexOf(name) === -1) libs.push(name); }
-
-        // 1. Window globals (works for CDN-loaded / non-bundled libraries)
-        if (typeof window.gsap !== 'undefined' || typeof window.TweenMax !== 'undefined') add('GSAP');
-        if (typeof window.ScrollTrigger !== 'undefined') add('GSAP ScrollTrigger');
-        if (typeof window.THREE !== 'undefined') add('Three.js');
-        if (typeof window.PIXI !== 'undefined') add('PixiJS');
-        if (typeof window.BABYLON !== 'undefined') add('Babylon.js');
-        if (typeof window.Lottie !== 'undefined' || typeof window.lottie !== 'undefined') add('Lottie');
-        if (typeof window.__NEXT_DATA__ !== 'undefined') add('Next.js');
-        if (typeof window.__NUXT__ !== 'undefined') add('Nuxt');
-        if (typeof window.Webflow !== 'undefined') add('Webflow');
-
-        // 2. DOM fingerprints (survive bundling — most reliable for modern sites)
-        // Three.js sets data-engine on every canvas it creates
-        var threeCanvas = document.querySelector('canvas[data-engine*="three"]');
-        if (threeCanvas) add('Three.js (' + (threeCanvas.getAttribute('data-engine') || '') + ')');
-        // Babylon.js also sets data-engine
-        var babylonCanvas = document.querySelector('canvas[data-engine*="Babylon"]');
-        if (babylonCanvas) add('Babylon.js');
-        // Lottie web components
-        if (document.querySelector('dotlottie-wc, lottie-player, dotlottie-player')) add('Lottie');
-        // Rive
-        if (document.querySelector('canvas[class*="rive"], rive-canvas')) add('Rive');
-        // React/Next.js
-        if (document.getElementById('__next')) add('Next.js');
-        if (document.getElementById('__nuxt')) add('Nuxt');
-        if (document.querySelector('[data-reactroot], [data-react-helmet]')) add('React');
-        // Svelte
-        if (document.querySelector('[class*="svelte-"]')) add('Svelte');
-        // Tailwind (utility class detection)
-        if (document.querySelector('[class*="flex "], [class*="grid "], [class*="px-"], [class*="py-"]')) add('Tailwind CSS');
-        // Framer Motion
-        if (document.querySelector('[style*="--framer-"], [data-framer-component-type]')) add('Framer Motion');
-
-        // 3. Script URL patterns
-        document.querySelectorAll('script[src]').forEach(function(s) {
-          var src = s.src.toLowerCase();
-          if (src.includes('gsap') || src.includes('tweenmax') || src.includes('greensock')) add('GSAP');
-          if (src.includes('scrolltrigger')) add('GSAP ScrollTrigger');
-          if (src.includes('three.module') || src.includes('three.min')) add('Three.js');
-          if (src.includes('pixi')) add('PixiJS');
-          if (src.includes('lottie') || src.includes('bodymovin')) add('Lottie');
-          if (src.includes('framer-motion')) add('Framer Motion');
-          if (src.includes('anime.min') || src.includes('animejs')) add('Anime.js');
-          if (src.includes('matter.min') || src.includes('matter-js')) add('Matter.js');
-          if (src.includes('lenis')) add('Lenis (smooth scroll)');
-        });
-
-        return libs;
-      })()`)) as string[];
-    } catch {
-      // Non-blocking
-    }
-
-    // 4. Shader fingerprinting — infer WebGL framework from captured GLSL
-    try {
-      const capturedShaders = await page1.evaluate(`window.__capturedShaders || []`);
-      if (Array.isArray(capturedShaders) && capturedShaders.length > 0) {
-        const allSource = (capturedShaders as Array<{ source: string }>)
-          .map((s) => s.source)
-          .join("\n");
-        const add = (name: string) => {
-          if (!detectedLibraries.includes(name)) detectedLibraries.push(name);
-        };
-        add("WebGL");
-        // Three.js shader fingerprints (built-in uniforms that survive bundling)
-        if (allSource.includes("modelViewMatrix") && allSource.includes("projectionMatrix"))
-          add("Three.js (confirmed via shaders)");
-        // PixiJS shader fingerprints
-        else if (
-          allSource.includes("vTextureCoord") &&
-          allSource.includes("uSampler") &&
-          !allSource.includes("modelViewMatrix")
-        )
-          add("PixiJS (confirmed via shaders)");
-        // Babylon.js shader fingerprints
-        else if (allSource.includes("viewProjection") && allSource.includes("world"))
-          add("Babylon.js (confirmed via shaders)");
-      }
-    } catch {
-      /* non-blocking */
-    }
+    // Detect JS libraries via globals, DOM fingerprints, script URLs, and shaders
+    const detectedLibraries = await detectLibraries(page1, capturedShaders);
 
     // Extract all visible text in DOM order
-    let visibleTextContent = "";
-    try {
-      visibleTextContent = (await page1.evaluate(`(() => {
-        var walker = document.createTreeWalker(document.body, NodeFilter.SHOW_TEXT, null);
-        var texts = [];
-        var node;
-        while (node = walker.nextNode()) {
-          var text = (node.textContent || '').trim();
-          if (text.length < 3) continue;
-          var el = node.parentElement;
-          if (!el) continue;
-          var style = getComputedStyle(el);
-          if (style.display === 'none' || style.visibility === 'hidden' || style.opacity === '0') continue;
-          var tag = el.tagName.toLowerCase();
-          if (tag === 'script' || tag === 'style' || tag === 'noscript') continue;
-          texts.push(text);
-        }
-        return texts.join('\\n');
-      })()`)) as string;
-      // Truncate to ~30K chars to avoid blowing up the prompt
-      if (visibleTextContent.length > 30000) {
-        visibleTextContent = visibleTextContent.slice(0, 30000) + "\n[...truncated]";
-      }
-    } catch {
-      // Non-blocking
-    }
+    const visibleTextContent = await extractVisibleText(page1);
 
     await page1.close();
 
@@ -868,151 +459,12 @@ a.addEventListener('DOMLoaded',function(){a.goToAndStop(${midFrame},true);window
     }
 
     // AI-powered image captioning via Gemini (optional — enriches asset descriptions)
-    const geminiCaptions: Record<string, string> = {};
-    const geminiKey = process.env.GEMINI_API_KEY || process.env.GOOGLE_API_KEY;
-    if (geminiKey) {
-      progress("design", "Captioning images with Gemini vision...");
-      try {
-        const { GoogleGenAI } = await import("@google/genai");
-        const { readFileSync: readAssetFile, readdirSync: listAssetDir } = await import("node:fs");
-        const ai = new GoogleGenAI({ apiKey: geminiKey });
-        const imageFiles = listAssetDir(join(outputDir, "assets")).filter((f: string) =>
-          /\.(png|jpg|jpeg|webp|gif)$/i.test(f),
-        );
-
-        // Caption in parallel batches via Gemini vision API.
-        // Rate limits vary by tier: free = 5 RPM, paid = 1000+ RPM.
-        // We batch 5 at a time with a 12s pause — safe for free tier, fast for paid.
-        const model = "gemini-2.5-flash";
-        const BATCH_SIZE = 5;
-        for (let i = 0; i < imageFiles.length; i += BATCH_SIZE) {
-          const batch = imageFiles.slice(i, i + BATCH_SIZE);
-          const results = await Promise.allSettled(
-            batch.map(async (file: string) => {
-              const filePath = join(outputDir, "assets", file);
-              const buffer = readAssetFile(filePath);
-              const base64 = buffer.toString("base64");
-              const ext = file.split(".").pop()?.toLowerCase() || "png";
-              const mimeType = ext === "jpg" ? "image/jpeg" : `image/${ext}`;
-              const response = await ai.models.generateContent({
-                model,
-                contents: [
-                  {
-                    role: "user",
-                    parts: [
-                      { inlineData: { mimeType, data: base64 } },
-                      {
-                        text: "Describe this website image in ONE short sentence for a video storyboard. Focus on: what it shows, dominant colors, whether background is light or dark. Be factual, not creative.",
-                      },
-                    ],
-                  },
-                ],
-                config: { maxOutputTokens: 300 },
-              });
-              return { file, caption: response.text?.trim() || "" };
-            }),
-          );
-          for (const result of results) {
-            if (result.status === "fulfilled" && result.value.caption) {
-              geminiCaptions[result.value.file] = result.value.caption;
-            }
-          }
-          // Pace requests to stay under free tier rate limits (5 RPM for gemini-2.5-flash)
-          if (i + BATCH_SIZE < imageFiles.length) {
-            await new Promise((r) => setTimeout(r, 12000)); // 12s between batches of 5 ≈ 25 RPM (safe for free tier)
-          }
-          progress(
-            "design",
-            `Captioned ${Math.min(i + BATCH_SIZE, imageFiles.length)}/${imageFiles.length} images...`,
-          );
-        }
-        progress("design", `${Object.keys(geminiCaptions).length} images captioned with Gemini`);
-      } catch (err) {
-        warnings.push(`Gemini captioning failed: ${err}`);
-      }
-    }
+    const geminiCaptions = await captionImagesWithGemini(outputDir, progress, warnings);
 
     // Generate asset descriptions for the AI agent
     progress("design", "Generating asset descriptions...");
     try {
-      const { readdirSync, statSync } = await import("node:fs");
-      const lines: string[] = [];
-
-      // Describe downloaded images
-      const assetsPath = join(outputDir, "assets");
-      try {
-        for (const file of readdirSync(assetsPath)) {
-          if (file === "svgs" || file === "fonts" || file === "lottie" || file === "videos")
-            continue;
-          const filePath = join(assetsPath, file);
-          const stat = statSync(filePath);
-          if (!stat.isFile()) continue;
-          const sizeKb = Math.round(stat.size / 1024);
-          // Find context from cataloged assets
-          const catalogMatch = catalogedAssets.find(
-            (a) =>
-              a.url && file.includes(a.url.split("/").pop()?.split("?")[0]?.slice(0, 20) || "___"),
-          );
-          // Build rich description from catalog context
-          const desc = catalogMatch?.description || catalogMatch?.notes || "";
-          const heading = catalogMatch?.nearestHeading || "";
-          const section = catalogMatch?.sectionClasses || "";
-          const aboveFold = catalogMatch?.aboveFold ? "above fold" : "";
-          // No brightness analysis — Gemini captions handle this when available
-          // Gemini vision caption (highest quality — available when GEMINI_API_KEY is set)
-          const geminiCaption = geminiCaptions[file];
-          // Derive a meaningful name from the filename
-          const cleanName = file.replace(/\.[^.]+$/, "").replace(/[-_]/g, " ");
-          const parts = [`${file} — ${sizeKb}KB`];
-          if (geminiCaption) {
-            // Gemini gives the best description — use it as primary
-            parts.push(geminiCaption);
-          } else {
-            // Fallback: DOM context
-            if (desc) parts.push(`"${desc.slice(0, 80)}"`);
-            if (heading) parts.push(`section: "${heading.slice(0, 60)}"`);
-            else if (section) parts.push(`in: ${section.split(" ").slice(0, 3).join(" ")}`);
-            if (aboveFold) parts.push(aboveFold);
-            if (!desc && !heading) parts.push(cleanName);
-          }
-          lines.push(parts.join(", "));
-        }
-      } catch {
-        /* no assets dir */
-      }
-
-      // Describe SVGs
-      try {
-        const svgsPath = join(assetsPath, "svgs");
-        for (const file of readdirSync(svgsPath)) {
-          if (!file.endsWith(".svg")) continue;
-          const svgMatch = tokens.svgs.find(
-            (s) =>
-              s.label &&
-              file.includes(
-                s.label
-                  .toLowerCase()
-                  .replace(/[^a-z0-9]/g, "-")
-                  .slice(0, 15),
-              ),
-          );
-          const label = svgMatch?.label || file.replace(".svg", "").replace(/-/g, " ");
-          const isLogo = svgMatch?.isLogo || file.includes("logo");
-          lines.push(`svgs/${file} — ${isLogo ? "logo: " : "icon: "}${label}`);
-        }
-      } catch {
-        /* no svgs dir */
-      }
-
-      // Describe fonts
-      try {
-        const fontsPath = join(assetsPath, "fonts");
-        for (const file of readdirSync(fontsPath)) {
-          lines.push(`fonts/${file} — font file`);
-        }
-      } catch {
-        /* no fonts dir */
-      }
+      const lines = generateAssetDescriptions(outputDir, tokens, catalogedAssets, geminiCaptions);
 
       if (lines.length > 0) {
         writeFileSync(
@@ -1080,77 +532,19 @@ a.addEventListener('DOMLoaded',function(){a.goToAndStop(${midFrame},true);window
       }
     }
 
-    // Ensure capture output is a valid HyperFrames project (index.html + meta.json)
-    const indexPath = join(outputDir, "index.html");
-    const metaPath = join(outputDir, "meta.json");
-    if (!existsSync(indexPath)) {
-      writeFileSync(
-        indexPath,
-        `<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=1920, height=1080" />
-    <script src="https://cdn.jsdelivr.net/npm/gsap@3.14.2/dist/gsap.min.js"></script>
-    <style>
-      * { margin: 0; padding: 0; box-sizing: border-box; }
-      html, body { margin: 0; width: 1920px; height: 1080px; overflow: hidden; background: #000; }
-    </style>
-  </head>
-  <body>
-    <!-- Root composition wrapper — AGENT: update data-duration to match total video length -->
-    <div data-composition-id="main" data-width="1920" data-height="1080" data-start="0" data-duration="28">
-
-      <!-- SCENE SLOTS — AGENT: adjust count, durations, and IDs to match your scene plan -->
-      <div id="scene-1" data-composition-src="compositions/scene-1.html" data-start="0" data-duration="7" data-track-index="1" data-width="1920" data-height="1080"></div>
-      <div id="scene-2" data-composition-src="compositions/scene-2.html" data-start="7" data-duration="7" data-track-index="1" data-width="1920" data-height="1080"></div>
-      <div id="scene-3" data-composition-src="compositions/scene-3.html" data-start="14" data-duration="7" data-track-index="1" data-width="1920" data-height="1080"></div>
-      <div id="scene-4" data-composition-src="compositions/scene-4.html" data-start="21" data-duration="7" data-track-index="1" data-width="1920" data-height="1080"></div>
-
-      <!-- NARRATION — AGENT: update src after generating TTS -->
-      <audio id="narration" data-start="0" data-duration="28" data-track-index="0" data-volume="1" src="narration.wav"></audio>
-
-      <!-- CAPTIONS (optional — only add if user requests captions/subtitles) -->
-
-    </div>
-
-    <script>
-      window.__timelines = window.__timelines || {};
-      var tl = gsap.timeline({ paused: true });
-      window.__timelines["main"] = tl;
-    </script>
-  </body>
-</html>
-`,
-        "utf-8",
-      );
-    }
-    if (!existsSync(metaPath)) {
-      const hostname = new URL(url).hostname.replace(/^www\./, "");
-      writeFileSync(
-        metaPath,
-        JSON.stringify({ id: hostname + "-video", name: tokens.title || hostname }, null, 2),
-        "utf-8",
-      );
-    }
-
-    // Generate CLAUDE.md + .cursorrules (AI agent instructions — always, regardless of API keys)
-    try {
-      const { generateAgentPrompt } = await import("./agentPromptGenerator.js");
-      generateAgentPrompt(
-        outputDir,
-        url,
-        tokens,
-        animationCatalog,
-        screenshots.length > 0,
-        discoveredLotties.length > 0,
-        existsSync(join(outputDir, "extracted", "shaders.json")),
-        catalogedAssets,
-      );
-      progress("agent", "CLAUDE.md generated");
-    } catch (err) {
-      warnings.push(`CLAUDE.md generation failed: ${err}`);
-    }
+    // Generate project scaffold (index.html, meta.json, CLAUDE.md)
+    await generateProjectScaffold(
+      outputDir,
+      url,
+      tokens,
+      animationCatalog,
+      screenshots.length > 0,
+      discoveredLotties.length > 0,
+      existsSync(join(outputDir, "extracted", "shaders.json")),
+      catalogedAssets,
+      progress,
+      warnings,
+    );
 
     progress("done", "Capture complete");
